refactor eth deposit poller
