+++ services/claim-relayer/utils/startEthDepositPoller.ts
@@ -9,11 +9,11 @@ import {
 } from "@claim-relayer/libs/constants";
 import { getRabbitMQSet } from "@bs-libs/utils/getRabbitMQSet";
 import * as ERC20Peg from "@bs-libs/abi/ERC20Peg.json";
-import { ethers, Event } from "ethers";
-import { handleDepositEvent } from "@claim-relayer/utils/handleDepositEvent";
-import { BridgeClaim } from "@claim-relayer/libs/models";
+import { ethers } from "ethers";
 import { u64 } from "@cennznet/types";
 import { waitFor } from "@bs-libs/utils/waitFor";
+import { handleMissedEvents } from "@claim-relayer/utils/handleMissedEvents";
+import { getMissedDepositEvents } from "@claim-relayer/utils/getMissedDepositEvents";
 
 const { PEG_CONTRACT_ADDRESS } = NETWORK_DETAILS;
 
@@ -41,45 +41,21 @@ export async function startEthDepositPoller(
 			logger.error("Mongo DB disconnected Poller stopping...");
 			break;
 		}
+
 		const allEvents = await peg.queryFilter({});
-		const depositEventTxHashes = allEvents
-			.filter((event) => event.event === "Deposit")
-			.map((event) => event.transactionHash);
-		//check if any deposit tx hashes are not in DB yet
-		const allDepositClaims = await BridgeClaim.find({});
-		const allDepositClaimsTxHashes = allDepositClaims.map(
-			(claim) => claim.txHash
-		);
-		const missedDepositEventHashes = depositEventTxHashes.filter(
-			(txHash) => !allDepositClaimsTxHashes.includes(txHash)
-		);
-		logger.info(
-			`Current Missed Deposit Events Number: ${missedDepositEventHashes?.length}`
-		);
-		//get the event for each and submit
+
+		const missedDepositEventHashes = await getMissedDepositEvents(allEvents);
 		const eventConfirmations = (
 			(await cennzApi.query.ethBridge.eventConfirmations()) as u64
 		).toNumber();
-		const missedEventProms = missedDepositEventHashes.map(async (txHash) => {
-			const eventInfo = allEvents.find(
-				(event) => txHash === event.transactionHash
-			) as Event;
 
-			const cennznetAddress = (eventInfo as any).args[3];
-			const amount = (eventInfo as any).args[2].toString();
-			const tokenAddress = (eventInfo as any).args[1];
-			await handleDepositEvent(
-				cennzApi,
-				eventInfo.transactionHash,
-				cennznetAddress,
-				amount,
-				tokenAddress,
-				eventConfirmations,
-				queue,
-				logger
-			);
-		});
-		await Promise.all(missedEventProms);
+		await handleMissedEvents(
+			cennzApi,
+			queue,
+			allEvents,
+			missedDepositEventHashes,
+			eventConfirmations
+		);
 		await waitFor(ETH_DEPOST_POLLER_INTERVAL);
 	}
 }
